Tried to move over to size_t enc for all things length related.

+++ crypto/include/aes_gcm.h
@@ -56,8 +56,8 @@
 #include <openssl/aes.h>
 
 typedef struct {
-    int key_size;
-    int tag_len;
+    size_t key_size;
+    size_t tag_len;
     EVP_CIPHER_CTX *ctx;
     srtp_cipher_direction_t dir;
 } srtp_aes_gcm_ctx_t;
