+++ crypto/include/aes_icm_ext.h
@@ -57,7 +57,7 @@
 typedef struct {
     v128_t counter; /* holds the counter value          */
     v128_t offset;  /* initial offset value             */
-    int key_size;
+    size_t key_size;
     EVP_CIPHER_CTX *ctx;
 } srtp_aes_icm_ctx_t;
