+++ crypto/replay/rdb.c
@@ -116,7 +116,7 @@ srtp_err_status_t srtp_rdb_add_index(srtp_rdb_t *rdb, uint32_t p_index)
         delta -= rdb_bits_in_bitmask - 1;
 
         /* shift the window forward by delta bits*/
-        v128_left_shift(&rdb->bitmask, delta);
+        v128_left_shift(&rdb->bitmask, (int)delta);
         v128_set_bit(&rdb->bitmask, rdb_bits_in_bitmask - 1);
         rdb->window_start += delta;
     }
